+++ rom.py
@@ -221,6 +221,7 @@ def generate_output(world: "PokemonPlatinumWorld", output_directory: str, patch:
     add_opt_byte("hm_cut_ins")
     add_opt_byte("buck_pos")
     ap_bin += (world.options.hb_speed.value - 1).to_bytes(length=1, byteorder='little')
+    add_opt_byte("normalize_encounters")
 
     if len(ap_bin) % 2 == 1:
         ap_bin += b'\x00'